pinstream/backend/pin_stream/requirements.txt
Flask==0.11.1
Flask-SQLAlchemy==2.1
flask_httpauth==3.2.3
mysql-python==1.2.5